rename new.rs to io.rs
In the long ago, "new" referred to the "new I/O module" after I/O reform
in the pre-1.0 days.
